% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{fishery}
\alias{fishery}
\title{Fishery data.}
\format{
A data frame with 677 rows and 2 variables
}
\usage{
data(fishery)
}
\description{
The fishery data consist of 677 transactions of a fishery product in Europe.
 For each transaction the Value in 1000 euro and the quantity in Tons are reported.
}
\keyword{datasets}
