% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{flea}
\alias{flea}
\title{Flea}
\format{
A data frame with 74 rows and 7 variables: six explanatory and one response variable - \code{species}.
The variables are as follows:

\itemize{
  \item tars1: width of the first joint of the first tarsus in microns (the sum of measurements for both tarsi)
  \item tars2: the same for the second joint
  \item head: the maximal width of the head between the external edges of the eyes in 0.01 mm
  \item ade1: the maximal width of the aedeagus in the fore-part in microns
  \item ade2: the front angle of the aedeagus ( 1 unit = 7.5 degrees)
  \item ade3: the aedeagus width from the side in microns
  \item species, which species is being examined - \code{Concinna}, \code{Heptapotamica}, \code{Heikertingeri}
}
}
\usage{
data(flea)
}
\description{
Flea-beetle measurements
}
\examples{
 data(flea)
 head(flea)

}
\references{
A. A. Lubischew (1962), On the Use of Discriminant Functions in Taxonomy, \emph{Biometrics}, \bold{18}4 pp.455--477.
}
\keyword{datasets}
