% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{Hospital data (Neter et al., 1996)}
\format{
A data frame with 108 rows and 5 variables
The variables are as follows:

\itemize{
  \item X1
  \item X2
  \item X3
  \item X4
  \item y
}

 @source
 J. NETER, M. H. KUTNER, C. J. NACHTSHEIM, W.WASSERMAN, \emph{Applied Linear Statistical Models} (4th edition). McGraw-Hill, New York, 1996.

 @references
 A. C. ATKINSON, M. RIANI, \emph{Robust Diagnostic Regression Analysis}. Springer-Verlag, New York, 2000.
}
\usage{
data(hospital)
}
\description{
Data on the logged survival time of 108 patients undergoing liver surgery,
 together with four potential explanatory variables. Data are composed of
 54 observations plus other 54 observations, introduced to check the model
 fitted to the first 54. Their comparison suggests there is no systematic
 difference between the two sets. However by looking at some FS plots
 (Riani and Atkinson, 2007), we conclude that these two groups are significantly
 different
}
\keyword{datasets}
