% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Income1}
\alias{Income1}
\title{Income1}
\format{
A data frame with 200 rows and 4 variables.
The variables are as follows:

\itemize{
  \item H_NUMPER: Number of persons in household
  \item HOTHVAL: All other types of income except HEARNVAL Recode - Total other household income
  \item HSSVAL: household income - social security
  \item HTOTVAL: total household income (dollar amount)
}
}
\source{
United States Census Bureau (2021). Annual Social and Economic Supplements
}
\usage{
data(Income1)
}
\description{
Income data taken from the United States Census Bureau. The data are
 a random sample of 200 observations referred to four variables.
 The goal is to predict HTOTVAL.
}
\examples{
 data(Income1)
 head(Income1)

}
\keyword{datasets}
