% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Income2}
\alias{Income2}
\title{Income2}
\format{
A data frame with 200 rows and 6 variables.
The variables are as follows:

\itemize{
 \item Age:  Age of the person (the minimum is 19 and the maximum is 73).
 \item Education: Number of years of education (the minimum value of 5 is
     primary school, and the maximum value is 16 bachelor degree)
 \item Gender: A factor - Male or Female
 \item ExtraGain: Income from investment sources (profit-losses)
     apart from wages/salary
 \item Hours: total number of declared hours worked during the week.
     The minimum value is 35 and the maximum is 99
 \item Income: total yearly income (Euro amount)
}
}
\usage{
data(Income2)
}
\description{
A sample of 200 observations of full time employees from a municipality
 in Northern Italy who have declared extra income from investment
 sources. The variables are as follows.
 The goal is the possibility in predicting income level based on
 the individual's personal information.
}
\examples{
 data(Income2)
 head(Income2)

}
\keyword{datasets}
