% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{wool}
\alias{wool}
\title{Wool data.}
\format{
A data frame with 27 rows and 4 variables
}
\usage{
data(wool)
}
\description{
The wool data give the number of cycles to failure of a worsted yarn under
 cycles of repeated loading. The variables are: length of test specimen;
 amplitude of loading cycle; load
}
\keyword{datasets}
