% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{z1}
\alias{z1}
\title{z1}
\format{
A data frame with 150 rows and 2 variables.
The variables are as follows:

\itemize{
  \item X1
  \item X2
}
}
\usage{
data(z1)
}
\description{
Simulated data to test tclustIC() and tclustICsol(), carbike() functions
}
\examples{
 data(z1)
 head(z1)
 \dontrun{
 (out <- tclustIC(z1, plots=FALSE, whichIC="CLACLA"))
 (outCLACLA <- tclustICsol(out, whichIC="CLACLA", plot=FALSE))
 carbikeplot(outCLACLA)
 }

}
\references{
Maitra, R. and Melnykov, V. (2010), Simulating data to study performance
 of finite mixture modeling and clustering algorithms, \emph{The Journal of
 Computational and Graphical Statistics}, Vol. 19, pp. 354-376.
}
\keyword{datasets}
